refactor(root): root controller 분리
