+++ src/main/java/com/dduru/gildongmu/home/controller/HomeApiDocs.java
@@ -1,17 +1,7 @@
 package com.dduru.gildongmu.home.controller;
 
-import com.dduru.gildongmu.common.dto.ApiResult;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import org.springframework.http.ResponseEntity;
-
-import java.util.Map;
 
 @Tag(name = "Home", description = "홈 API")
 public interface HomeApiDocs {
-
-    @Operation(summary = "웰컴 메시지", description = "서비스 기본 정보를 json형식으로 반환합니다.")
-    @ApiResponse(responseCode = "200", description = "성공")
-    ResponseEntity<ApiResult<Map<String, Object>>> welcome();
 }
